# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ccce-atlas API",
        "description": "Civic geospatial platform API for Corpus Christi / Nueces County, Texas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/parcels/": {
            "get": {
                "tags": [
                    "parcels"
                ],
                "summary": "List Parcels",
                "description": "List parcels with optional filters.\n\nFilters:\n- owner: Partial match (case-insensitive)\n- min_value/max_value: Appraised value range\n- class_cd: Property class code\n- bbox: Bounding box (west,south,east,north)\n\nReturns:\n    Paginated list of parcels",
                "operationId": "list_parcels_api_v1_parcels__get",
                "parameters": [
                    {
                        "name": "owner",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Owner name (case-insensitive partial match)",
                            "title": "Owner"
                        },
                        "description": "Owner name (case-insensitive partial match)"
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Minimum appraised value",
                            "title": "Min Value"
                        },
                        "description": "Minimum appraised value"
                    },
                    {
                        "name": "max_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Maximum appraised value",
                            "title": "Max Value"
                        },
                        "description": "Maximum appraised value"
                    },
                    {
                        "name": "class_cd",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Property class code",
                            "title": "Class Cd"
                        },
                        "description": "Property class code"
                    },
                    {
                        "name": "bbox",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Bounding box: west,south,east,north",
                            "title": "Bbox"
                        },
                        "description": "Bounding box: west,south,east,north"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "description": "Results per page",
                            "default": 100,
                            "title": "Limit"
                        },
                        "description": "Results per page"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Offset for pagination",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "Offset for pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParcelList"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/parcels/spatial": {
            "post": {
                "tags": [
                    "parcels"
                ],
                "summary": "Find Intersecting Parcels",
                "description": "Find parcels intersecting a GeoJSON geometry (e.g., user-drawn polygon).\n\nRequest body:\n    {\n        \"geometry\": {...},  // GeoJSON geometry\n        \"limit\": 100\n    }\n\nReturns:\n    Parcels intersecting the geometry, sorted by intersection area",
                "operationId": "find_intersecting_parcels_api_v1_parcels_spatial_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpatialIntersectionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpatialIntersectionResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/pois/": {
            "get": {
                "tags": [
                    "pois"
                ],
                "summary": "List Pois",
                "description": "List POIs with optional filters.\n\nFilters:\n- category: POI category (beaches, trails, eats, coffee, etc.)\n- bbox: Bounding box (west,south,east,north)\n\nReturns:\n    Paginated list of POIs",
                "operationId": "list_pois_api_v1_pois__get",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "POI category filter",
                            "title": "Category"
                        },
                        "description": "POI category filter"
                    },
                    {
                        "name": "bbox",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Bounding box: west,south,east,north",
                            "title": "Bbox"
                        },
                        "description": "Bounding box: west,south,east,north"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "description": "Results per page",
                            "default": 100,
                            "title": "Limit"
                        },
                        "description": "Results per page"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Offset for pagination",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "Offset for pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/POIList"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "POI": {
                "properties": {
                    "poi_id": {
                        "type": "string",
                        "title": "Poi Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "category": {
                        "type": "string",
                        "title": "Category"
                    },
                    "subcategory": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Subcategory"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address"
                    },
                    "phone": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phone"
                    },
                    "website": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Website"
                    },
                    "hours": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hours"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "geometry": {
                        "title": "Geometry"
                    },
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "poi_id",
                    "name",
                    "category",
                    "id",
                    "geometry"
                ],
                "title": "POI",
                "description": "Full POI response with geometry."
            },
            "POIList": {
                "properties": {
                    "pois": {
                        "items": {
                            "$ref": "#/components/schemas/POI"
                        },
                        "type": "array",
                        "title": "Pois"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "limit": {
                        "type": "integer",
                        "title": "Limit"
                    },
                    "offset": {
                        "type": "integer",
                        "title": "Offset"
                    }
                },
                "type": "object",
                "required": [
                    "pois",
                    "total",
                    "limit",
                    "offset"
                ],
                "title": "POIList",
                "description": "Paginated list of POIs."
            },
...
            "Parcel": {
                "properties": {
                    "parcel_id": {
                        "type": "string",
                        "title": "Parcel Id"
                    },
                    "owner": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Owner"
                    },
                    "prop_addr": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prop Addr"
                    },
                    "zip_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Zip Code"
                    },
                    "appraised_value": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Appraised Value"
                    },
                    "market_value": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Market Value"
                    },
                    "land_acres": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Land Acres"
                    },
                    "class_cd": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Class Cd"
                    },
                    "year_built": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Year Built"
                    },
                    "zoning": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Zoning"
                    },
                    "prop_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prop Type"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "objectid": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Objectid"
                    },
                    "geometry": {
                        "title": "Geometry"
                    },
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "parcel_id",
                    "id",
                    "geometry"
                ],
                "title": "Parcel",
                "description": "Full parcel response with geometry."
            },
...
            "ParcelList": {
                "properties": {
                    "parcels": {
                        "items": {
                            "$ref": "#/components/schemas/Parcel"
                        },
                        "type": "array",
                        "title": "Parcels"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "limit": {
                        "type": "integer",
                        "title": "Limit"
                    },
                    "offset": {
                        "type": "integer",
                        "title": "Offset"
                    }
                },
                "type": "object",
                "required": [
                    "parcels",
                    "total",
                    "limit",
                    "offset"
                ],
                "title": "ParcelList",
                "description": "Paginated list of parcels."
            },
...
            "SpatialIntersectionRequest": {
                "properties": {
                    "geometry": {
                        "type": "object",
                        "title": "Geometry"
                    },
                    "limit": {
                        "type": "integer",
                        "title": "Limit",
                        "default": 100
                    }
                },
                "type": "object",
                "required": [
                    "geometry"
                ],
                "title": "SpatialIntersectionRequest",
                "description": "Request body for spatial intersection query."
            },
            "SpatialIntersectionResult": {
                "properties": {
                    "parcels": {
                        "items": {
                            "$ref": "#/components/schemas/ParcelIntersection"
                        },
                        "type": "array",
                        "title": "Parcels"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "parcels",
                    "total"
                ],
                "title": "SpatialIntersectionResult",
                "description": "Parcels intersecting a geometry."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}